requirements.txt
streamlit
psycopg2-binary        # For PostgreSQL connection
llama_index            # For managing embeddings
transformers           # Hugging Face model for embedding generation
sqlalchemy    
langchain-cohere